BAEL-4230 refactoring project
