+++ persistence-modules/spring-data-jdbc/pom.xml
@@ -6,6 +6,7 @@
 		<groupId>com.baeldung</groupId>
 		<artifactId>parent-boot-2</artifactId>
 		<version>0.0.1-SNAPSHOT</version>
+		<relativePath>../../parent-boot-2</relativePath>
 	</parent>
 	<artifactId>spring-data-jdbc</artifactId>
 	<name>spring-data-jdbc</name>
