+++ persistence-modules/spring-data-jdbc/src/main/java/com/baeldung/springdatajdbcintro/DatabaseSeeder.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.springdatajdbcintro;
 
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
