+++ persistence-modules/spring-data-jdbc/src/main/java/com/baeldung/springdatajdbcintro/entity/Person.java
@@ -1,4 +1,4 @@
-package com.baeldung.entity;
+package com.baeldung.springdatajdbcintro.entity;
 
 import org.springframework.data.annotation.Id;
